added some multi-threading synchronization code
there are a few race-conditions involved with the volume up/down keys being
used to continually seek through music tracks.  in some cases the booleans
that are used to determine this logic are set in parallel resulting in an
invalid state that results in longpress messages being sent continually.

this code implements a series of mutex locks to prevent this from occurring